feat: Refactor liquidation-watcher endpoint in watcher.py
- Moved POST /liquidation-watcher endpoint from main app to router-based structure in watcher.py
